feat: refactor the architecture
put all folder in features directory
