+++ src/features/mail/mail.config.ts
