+++ src/features/mail/templates/confirmation.hbs
