+++ src/features/rights/right.model.ts
