+++ src/features/user-article/user-article.model.ts
@@ -1,6 +1,6 @@
 import { Column, Model, Table, ForeignKey } from 'sequelize-typescript';
-import { User } from '../users/user.model';
-import { Article } from '../articles/article.model';
+import { User } from 'src/features/users/user.model';
+import { Article } from 'src/features/articles/article.model';
 
 @Table({ tableName: 'user-article' })
 export class UserArticle extends Model {
