+++ src/features/user-team/user-team.model.ts
@@ -1,6 +1,6 @@
 import { Column, ForeignKey, Model, Table } from 'sequelize-typescript';
-import { User } from '../users/user.model';
-import { Team } from '../team/team.model';
+import { User } from 'src/features/users/user.model';
+import { Team } from 'src/features/team/team.model';
 
 @Table({ tableName: 'user-team' })
 export class UserTeam extends Model {
