--- test.txt
@@ -1,2 +1,0 @@
-ce document sert de test !
-test 2
